requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pyaes==1.6.1
pyasn1==0.5.0
python-dotenv==1.0.0
requests==2.30.0
rsa==4.9
soupsieve==2.4.1
Telethon==1.28.5
urllib3==2.0.2
